+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/clustering/ClusteringHelper.java
@@ -48,6 +48,7 @@ import org.preesm.algorithm.schedule.model.ParallelActorSchedule;
 import org.preesm.algorithm.schedule.model.ParallelSchedule;
 import org.preesm.algorithm.schedule.model.Schedule;
 import org.preesm.algorithm.schedule.model.SequentialSchedule;
+import org.preesm.algorithm.synthesis.schedule.iterator.SimpleScheduleIterator;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.model.pisdf.AbstractActor;
 import org.preesm.model.pisdf.AbstractVertex;
@@ -85,7 +86,7 @@ public class ClusteringHelper {
    */
   public static final List<DataInputPort> getExternalyConnectedPorts(final Schedule cluster) {
     List<DataInputPort> res = new ArrayList<>();
-    final List<AbstractActor> actors = cluster.getActors();
+    final List<AbstractActor> actors = new SimpleScheduleIterator(cluster).getOrderedList();
     for (final AbstractActor actor : actors) {
       final EList<DataInputPort> dataInputPorts = actor.getDataInputPorts();
       for (final DataInputPort port : dataInputPorts) {
@@ -201,7 +202,8 @@ public class ClusteringHelper {
       timing = getExecutionTimeOfHierarchical(schedule, scenario, component, timing);
     } else {
       // Retrieve timing from actors
-      AbstractActor actor = schedule.getActors().get(0);
+      final List<AbstractActor> actors = new SimpleScheduleIterator(schedule).getOrderedList();
+      AbstractActor actor = actors.get(0);
       long actorTiming = scenario.getTimings().evaluateTimingOrDefault(actor, component);
       if ((schedule instanceof ParallelActorSchedule)) {
         timing = actorTiming;