+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/synthesis/schedule/AbstractScheduler.java
@@ -40,6 +40,7 @@ import org.preesm.algorithm.mapping.model.Mapping;
 import org.preesm.algorithm.schedule.model.Schedule;
 import org.preesm.algorithm.synthesis.PreesmSynthesisException;
 import org.preesm.algorithm.synthesis.SynthesisResult;
+import org.preesm.algorithm.synthesis.schedule.iterator.SimpleScheduleIterator;
 import org.preesm.commons.model.PreesmCopyTracker;
 import org.preesm.model.pisdf.AbstractActor;
 import org.preesm.model.pisdf.EndActor;
@@ -101,7 +102,9 @@ public abstract class AbstractScheduler implements IScheduler {
 
     // make sure all actors have been scheduled and schedule contains only actors from the input graph
     final List<AbstractActor> piGraphAllActors = new ArrayList<>(piGraph.getAllActors());
-    final List<AbstractActor> scheduledActors = new ArrayList<>(schedule.getActors());
+
+    final List<AbstractActor> actors = new SimpleScheduleIterator(schedule).getOrderedList();
+    final List<AbstractActor> scheduledActors = new ArrayList<>(actors);
     if (!piGraphAllActors.containsAll(scheduledActors)) {
       throw new PreesmSynthesisException("Schedule refers actors not present in the input PiSDF.");
     }