+++ plugins/org.preesm.codegen/src/org/preesm/codegen/model/clustering/CodegenClusterModelGeneratorSwitch.java
@@ -49,6 +49,7 @@ import org.preesm.algorithm.schedule.model.ParallelHiearchicalSchedule;
 import org.preesm.algorithm.schedule.model.Schedule;
 import org.preesm.algorithm.schedule.model.SequentialHiearchicalSchedule;
 import org.preesm.algorithm.schedule.model.util.ScheduleSwitch;
+import org.preesm.algorithm.synthesis.schedule.iterator.SimpleScheduleIterator;
 import org.preesm.codegen.model.Block;
 import org.preesm.codegen.model.Buffer;
 import org.preesm.codegen.model.ClusterBlock;
@@ -242,7 +243,8 @@ public class CodegenClusterModelGeneratorSwitch extends ScheduleSwitch<CodeElt> 
 
     // Retrieve actor to fire
     // clustering process does list actors in actor schedule, we only care about the first one here
-    final AbstractActor actor = schedule.getActors().get(0);
+    final List<AbstractActor> actors = new SimpleScheduleIterator(schedule).getOrderedList();
+    final AbstractActor actor = actors.get(0);
 
     // Generate a LoopBlock to put function call element into
     final LoopBlock loopBlock = CodegenModelUserFactory.eINSTANCE.createLoopBlock();