+++ plugins/org.preesm.codegen/src/org/preesm/codegen/model/generator2/CodegenModelGenerator2.java
@@ -35,54 +35,55 @@
 package org.preesm.codegen.model.generator2;
 
 import java.util.Collections;
-import java.util.Deque;
 import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.logging.Level;
 import java.util.stream.Collectors;
-import org.apache.commons.collections4.BidiMap;
-import org.apache.commons.collections4.bidimap.DualHashBidiMap;
 import org.eclipse.emf.common.util.ECollections;
 import org.eclipse.emf.common.util.EList;
 import org.preesm.algorithm.mapping.model.Mapping;
 import org.preesm.algorithm.memalloc.model.Allocation;
-import org.preesm.algorithm.memalloc.model.DistributedBuffer;
-import org.preesm.algorithm.memalloc.model.LogicalBuffer;
 import org.preesm.algorithm.memalloc.model.PhysicalBuffer;
-import org.preesm.algorithm.memalloc.model.util.MemoryAllocationSwitch;
+import org.preesm.algorithm.schedule.model.CommunicationActor;
 import org.preesm.algorithm.schedule.model.Schedule;
+import org.preesm.algorithm.synthesis.schedule.iterator.ScheduleAndTopologyIterator;
 import org.preesm.codegen.model.ActorFunctionCall;
 import org.preesm.codegen.model.Block;
 import org.preesm.codegen.model.Buffer;
 import org.preesm.codegen.model.Call;
 import org.preesm.codegen.model.Constant;
 import org.preesm.codegen.model.CoreBlock;
+import org.preesm.codegen.model.FifoCall;
+import org.preesm.codegen.model.FifoOperation;
+import org.preesm.codegen.model.PortDirection;
 import org.preesm.codegen.model.SpecialCall;
 import org.preesm.codegen.model.SpecialType;
 import org.preesm.codegen.model.SubBuffer;
 import org.preesm.codegen.model.Variable;
 import org.preesm.codegen.model.util.CodegenModelUserFactory;
+import org.preesm.codegen.model.util.VariableSorter;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
+import org.preesm.commons.logger.PreesmLogger;
 import org.preesm.model.pisdf.AbstractActor;
 import org.preesm.model.pisdf.Actor;
 import org.preesm.model.pisdf.BroadcastActor;
 import org.preesm.model.pisdf.CHeaderRefinement;
-import org.preesm.model.pisdf.ConfigInputPort;
 import org.preesm.model.pisdf.EndActor;
 import org.preesm.model.pisdf.Fifo;
 import org.preesm.model.pisdf.ForkActor;
 import org.preesm.model.pisdf.FunctionPrototype;
-import org.preesm.model.pisdf.ISetter;
 import org.preesm.model.pisdf.InitActor;
 import org.preesm.model.pisdf.JoinActor;
-import org.preesm.model.pisdf.Parameter;
+import org.preesm.model.pisdf.PersistenceLevel;
 import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.pisdf.Port;
 import org.preesm.model.pisdf.Refinement;
 import org.preesm.model.pisdf.RoundBufferActor;
 import org.preesm.model.pisdf.SpecialActor;
+import org.preesm.model.pisdf.SrdagActor;
+import org.preesm.model.pisdf.UserSpecialActor;
 import org.preesm.model.scenario.Scenario;
 import org.preesm.model.slam.ComponentInstance;
 import org.preesm.model.slam.Design;
@@ -95,22 +96,20 @@ import org.preesm.model.slam.Design;
 public class CodegenModelGenerator2 {
 
   public static final List<Block> generate(final Design archi, final PiGraph algo, final Scenario scenario,
-      final Schedule schedule, final Mapping mapping, final Allocation memAlloc) {
-    return CodegenModelGenerator2.generate(archi, algo, scenario, schedule, mapping, memAlloc, false);
-  }
-
-  public static final List<Block> generate(final Design archi, final PiGraph algo, final Scenario scenario,
       final Schedule schedule, final Mapping mapping, final Allocation memAlloc, final boolean papify) {
     return new CodegenModelGenerator2(archi, algo, scenario, schedule, mapping, memAlloc, papify).generate();
   }
 
-  private final Design      archi;
-  private final PiGraph     algo;
-  private final Scenario    scenario;
-  private final Schedule    schedule;
-  private final Mapping     mapping;
-  private final Allocation  memAlloc;
-  private AllocationVisitor allocation;
+  private final Design     archi;
+  private final PiGraph    algo;
+  private final Scenario   scenario;
+  private final Schedule   schedule;
+  private final Mapping    mapping;
+  private final Allocation memAlloc;
+
+  private AllocationToCodegenBuffer memoryLinker;
+
+  private final boolean papify;
 
   private CodegenModelGenerator2(final Design archi, final PiGraph algo, final Scenario scenario,
       final Schedule schedule, final Mapping mapping, final Allocation memAlloc, final boolean papify) {
@@ -120,115 +119,12 @@ public class CodegenModelGenerator2 {
     this.schedule = schedule;
     this.mapping = mapping;
     this.memAlloc = memAlloc;
-  }
-
-  /**
-   *
-   */
-  class AllocationVisitor extends MemoryAllocationSwitch<Boolean> {
-
-    private final Deque<Buffer>                                     codegenBufferStack = new LinkedList<>();
-    private final Deque<org.preesm.algorithm.memalloc.model.Buffer> allocBufferStack   = new LinkedList<>();
-
-    final BidiMap<org.preesm.algorithm.memalloc.model.Buffer, Buffer> btb            = new DualHashBidiMap<>();
-    final Map<Port, Variable>                                         portToVariable = new LinkedHashMap<>();
-
-    @Override
-    public Boolean caseAllocation(final Allocation alloc) {
-      alloc.getPhysicalBuffers().forEach(this::doSwitch);
-      for (final Entry<Fifo, org.preesm.algorithm.memalloc.model.Buffer> allocations : alloc.getAllocations()) {
-        final Fifo fifo = allocations.getKey();
-        final org.preesm.algorithm.memalloc.model.Buffer buffer = allocations.getValue();
-        final Buffer codegenBuffer = this.btb.get(buffer);
-
-        codegenBuffer.setName(generateUniqueBufferName(fifo));
-        codegenBuffer.setType(fifo.getType());
-        codegenBuffer.setTypeSize(
-            CodegenModelGenerator2.this.scenario.getSimulationInfo().getDataTypeSizeOrDefault(fifo.getType()));
-
-        final String comment = fifo.getSourcePort().getContainingActor().getName() + " > "
-            + fifo.getTargetPort().getContainingActor().getName();
-        codegenBuffer.setComment(comment);
-
-        this.portToVariable.put(fifo.getTargetPort(), codegenBuffer);
-        this.portToVariable.put(fifo.getSourcePort(), codegenBuffer);
-      }
-      return true;
-    }
-
-    private final Map<String, Long> bufferNames = new LinkedHashMap<>();
-
-    private String generateUniqueBufferName(final Fifo fifo) {
-      final String candidate = fifo.getId().replace(".", "_").replace("-", "_");
-      long idx;
-      String key = candidate;
-      if (key.length() > 28) {
-        key = key.substring(0, 28);
-      }
-      if (this.bufferNames.containsKey(key)) {
-        idx = this.bufferNames.get(key);
-      } else {
-        idx = 0;
-        this.bufferNames.put(key, idx);
-      }
-
-      final String bufferName = key + "__" + idx;
-      idx += 1;
-      this.bufferNames.put(key, idx);
-      return bufferName;
-    }
-
-    @Override
-    public Boolean caseLogicalBuffer(final LogicalBuffer logicalBuffer) {
-      final SubBuffer subBuffer = CodegenModelUserFactory.eINSTANCE.createSubBuffer();
-      subBuffer.setSize(logicalBuffer.getSize());
-      subBuffer.setOffset(logicalBuffer.getOffset());
-
-      this.btb.put(logicalBuffer, subBuffer);
-      this.codegenBufferStack.push(subBuffer);
-      this.allocBufferStack.push(logicalBuffer);
-
-      final org.preesm.algorithm.memalloc.model.Buffer memory = logicalBuffer.getMemory();
-      final Buffer buffer = this.btb.get(memory);
-      subBuffer.reaffectContainer(buffer);
-
-      for (final org.preesm.algorithm.memalloc.model.Buffer l : logicalBuffer.getChildren()) {
-        doSwitch(l);
-      }
-
-      this.codegenBufferStack.pop();
-      this.allocBufferStack.pop();
-      return true;
-    }
-
-    @Override
-    public Boolean casePhysicalBuffer(final PhysicalBuffer phys) {
-      final Buffer mainBuffer = CodegenModelUserFactory.eINSTANCE.createBuffer();
-      mainBuffer.setSize(phys.getSize());
-      mainBuffer.setName("Shared_" + phys.getMemory().getHardwareId());
-      mainBuffer.setType("char");
-      mainBuffer.setTypeSize(1); // char is 1 byte
-      this.btb.put(phys, mainBuffer);
-      this.codegenBufferStack.push(mainBuffer);
-      this.allocBufferStack.push(phys);
-
-      for (final org.preesm.algorithm.memalloc.model.Buffer l : phys.getChildren()) {
-        doSwitch(l);
-      }
-
-      this.codegenBufferStack.pop();
-      this.allocBufferStack.pop();
-      return true;
-    }
-
-    @Override
-    public Boolean caseDistributedBuffer(final DistributedBuffer object) {
-      throw new UnsupportedOperationException();
-    }
-
+    this.papify = papify;
   }
 
   private List<Block> generate() {
+    final String msg = "Starting codegen2 with papify set to " + papify;
+    PreesmLogger.getLogger().log(Level.FINE, msg);
 
     final Map<ComponentInstance, CoreBlock> coreBlocks = new LinkedHashMap<>();
 
@@ -239,11 +135,11 @@ public class CodegenModelGenerator2 {
       coreBlocks.put(cmp, createCoreBlock);
     }
 
-    // 1- generate variables
-    final Map<Port, Variable> portToVariable = allocate();
+    // 1- generate variables (and keep track of them with a linker)
+    this.memoryLinker = AllocationToCodegenBuffer.link(memAlloc, scenario, algo);
 
     // 2- generate code
-    generateCode(coreBlocks, portToVariable);
+    generateCode(coreBlocks);
 
     // sort blocks
     final List<Block> resultList = coreBlocks.entrySet().stream()
@@ -251,172 +147,216 @@ public class CodegenModelGenerator2 {
         .collect(Collectors.toList());
 
     // generate buffer definitions
-    generateBuffers(portToVariable, coreBlocks);
+    generateBuffers(coreBlocks);
 
     return Collections.unmodifiableList(resultList);
   }
 
-  private void generateBuffers(final Map<Port, Variable> portToVariable,
-      final Map<ComponentInstance, CoreBlock> coreBlocks) {
-
-    for (final Entry<?, Buffer> entry : this.allocation.btb.entrySet()) {
-      final Buffer mainBuffer = entry.getValue();
-      final org.preesm.algorithm.memalloc.model.Buffer key = this.allocation.btb.getKey(mainBuffer);
-      final PhysicalBuffer memoryBankObj = key.getBank();
-      final String memoryBank = memoryBankObj.getMemory().getInstanceName();
-
-      // Identify the corresponding operator block.
-      // (also find out if the Buffer is local (i.e. not shared between
-      // several CoreBlock)
-      CoreBlock correspondingOperatorBlock = null;
-      final boolean isLocal;
-      final String correspondingOperatorID;
-
-      if (memoryBank.equalsIgnoreCase("shared_mem")) {
-        // If the memory bank is shared, let the main operator
-        // declare the Buffer.
-        correspondingOperatorID = this.scenario.getSimulationInfo().getMainOperator().getInstanceName();
-        isLocal = false;
-
-        // Check that the main operator block exists.
-        CoreBlock mainOperatorBlock = null;
-        for (final Entry<ComponentInstance, CoreBlock> componentEntry : coreBlocks.entrySet()) {
-          if (componentEntry.getKey().getInstanceName().equals(correspondingOperatorID)) {
-            mainOperatorBlock = componentEntry.getValue();
-          }
-        }
-
-        // If the main operator does not exist
-        if (mainOperatorBlock == null) {
-          // Create it
-          mainOperatorBlock = CodegenModelUserFactory.eINSTANCE.createCoreBlock(null);
-          final ComponentInstance componentInstance = this.archi.getComponentInstance(correspondingOperatorID);
-          mainOperatorBlock.setName(componentInstance.getInstanceName());
-          mainOperatorBlock.setCoreType(componentInstance.getComponent().getVlnv().getName());
-          coreBlocks.put(componentInstance, mainOperatorBlock);
-        }
-
-      } else {
-        // else, the operator corresponding to the memory bank will
-        // do the work
-        correspondingOperatorID = memoryBank;
-        isLocal = true;
-      }
-
-      // Find the block
-      for (final Entry<ComponentInstance, CoreBlock> componentEntry : coreBlocks.entrySet()) {
-        if (componentEntry.getKey().getInstanceName().equals(correspondingOperatorID)) {
-          correspondingOperatorBlock = componentEntry.getValue();
-        }
-      }
-      // Recursively set the creator for the current Buffer and all its
-      // subBuffer
-      recursiveSetBufferCreator(mainBuffer, correspondingOperatorBlock, isLocal);
-
-      if (correspondingOperatorBlock != null) {
-        final EList<Variable> definitions = correspondingOperatorBlock.getDefinitions();
-        ECollections.sort(definitions, (o1, o2) -> {
-          if ((o1 instanceof Buffer) && (o2 instanceof Buffer)) {
-            int sublevelO1 = 0;
-            if (o1 instanceof SubBuffer) {
-              Buffer b1 = (Buffer) o1;
-              while (b1 instanceof SubBuffer) {
-                sublevelO1++;
-                b1 = ((SubBuffer) b1).getContainer();
-              }
-            }
-
-            int sublevelO2 = 0;
-            if (o2 instanceof SubBuffer) {
-              Buffer b2 = (Buffer) o2;
-              while (b2 instanceof SubBuffer) {
-                sublevelO2++;
-                b2 = ((SubBuffer) b2).getContainer();
-              }
-            }
-
-            return sublevelO1 - sublevelO2;
-          }
-          if (o1 instanceof Buffer) {
-            return 1;
-          }
-          if (o2 instanceof Buffer) {
-            return -1;
-          }
-          return 0;
-        });
-      }
+  private void generateBuffers(final Map<ComponentInstance, CoreBlock> coreBlocks) {
+    final List<Buffer> buffers = this.memoryLinker.getCodegenBuffers();
+    for (Buffer buffer : buffers) {
+      generateBuffer(coreBlocks, buffer);
     }
   }
 
-  private void recursiveSetBufferCreator(final Variable buffer, final CoreBlock correspondingOperatorBlock,
+  private void generateBuffer(final Map<ComponentInstance, CoreBlock> coreBlocks, final Buffer mainBuffer) {
+    final org.preesm.algorithm.memalloc.model.Buffer key = this.memoryLinker.getAllocationBuffer(mainBuffer);
+    final PhysicalBuffer memoryBankObj = key.getBank();
+    final String memoryBank = memoryBankObj.getMemoryBank().getInstanceName();
+
+    // Identify the corresponding operator block.
+    // (also find out if the Buffer is local (i.e. not shared between
+    // several CoreBlock)
+    CoreBlock correspondingOperatorBlock = null;
+    final boolean isLocal;
+    final String correspondingOperatorID;
+
+    if (memoryBank.equalsIgnoreCase("shared_mem")) {
+      // If the memory bank is shared, let the main operator
+      // declare the Buffer.
+      correspondingOperatorID = this.scenario.getSimulationInfo().getMainOperator().getInstanceName();
+      isLocal = false;
+
+      // Check that the main operator block exists.
+      CoreBlock mainOperatorBlock = null;
+      for (final Entry<ComponentInstance, CoreBlock> componentEntry : coreBlocks.entrySet()) {
+        if (componentEntry.getKey().getInstanceName().equals(correspondingOperatorID)) {
+          mainOperatorBlock = componentEntry.getValue();
+        }
+      }
+
+      // If the main operator does not exist
+      if (mainOperatorBlock == null) {
+        // Create it
+        mainOperatorBlock = CodegenModelUserFactory.eINSTANCE.createCoreBlock(null);
+        final ComponentInstance componentInstance = this.archi.getComponentInstance(correspondingOperatorID);
+        mainOperatorBlock.setName(componentInstance.getInstanceName());
+        mainOperatorBlock.setCoreType(componentInstance.getComponent().getVlnv().getName());
+        coreBlocks.put(componentInstance, mainOperatorBlock);
+      }
+
+    } else {
+      // else, the operator corresponding to the memory bank will
+      // do the work
+      correspondingOperatorID = memoryBank;
+      isLocal = true;
+    }
+
+    // Find the block
+    for (final Entry<ComponentInstance, CoreBlock> componentEntry : coreBlocks.entrySet()) {
+      if (componentEntry.getKey().getInstanceName().equals(correspondingOperatorID)) {
+        correspondingOperatorBlock = componentEntry.getValue();
+      }
+    }
+    // Recursively set the creator for the current Buffer and all its
+    // subBuffer
+    recursiveSetBufferCreator(mainBuffer, correspondingOperatorBlock, isLocal);
+    sortDefinitions(correspondingOperatorBlock);
+  }
+
+  private void sortDefinitions(CoreBlock correspondingOperatorBlock) {
+    if (correspondingOperatorBlock != null) {
+      final EList<Variable> definitions = correspondingOperatorBlock.getDefinitions();
+      ECollections.sort(definitions, new VariableSorter());
+    }
+  }
+
+  private void recursiveSetBufferCreator(final Variable variable, final CoreBlock correspondingOperatorBlock,
       final boolean isLocal) {
     // Set the creator for the current buffer
-    buffer.reaffectCreator(correspondingOperatorBlock);
-    if (buffer instanceof Buffer) {
-      ((Buffer) buffer).setLocal(isLocal);
+    variable.reaffectCreator(correspondingOperatorBlock);
+    if (variable instanceof Buffer) {
+      final Buffer buffer = (Buffer) variable;
+      buffer.setLocal(isLocal);
       // Do the same recursively for all its children subbuffers
-      for (final SubBuffer subBuffer : ((Buffer) buffer).getChildrens()) {
+      for (final SubBuffer subBuffer : buffer.getChildrens()) {
         recursiveSetBufferCreator(subBuffer, correspondingOperatorBlock, isLocal);
       }
     }
   }
 
-  private void generateCode(final Map<ComponentInstance, CoreBlock> coreBlocks,
-      final Map<Port, Variable> portToVariable) {
+  private void generateCode(final Map<ComponentInstance, CoreBlock> coreBlocks) {
     // iterate in order
-    for (final AbstractActor actor : this.schedule.getActors()) {
+
+    final List<AbstractActor> actors = new ScheduleAndTopologyIterator(this.schedule).getOrderedList();
+    for (final AbstractActor actor : actors) {
       final EList<ComponentInstance> actorMapping = this.mapping.getMapping(actor);
       final ComponentInstance componentInstance = actorMapping.get(0);
       final CoreBlock coreBlock = coreBlocks.get(componentInstance);
 
       if (actor instanceof Actor) {
-        generateActorFiring((Actor) actor, portToVariable, coreBlock);
-      } else if (actor instanceof SpecialActor) {
-        generateSpecialActor((SpecialActor) actor, portToVariable, coreBlock);
+        generateActorFiring((Actor) actor, this.memoryLinker.getPortToVariableMap(), coreBlock);
+      } else if (actor instanceof UserSpecialActor) {
+        generateSpecialActor((UserSpecialActor) actor, this.memoryLinker.getPortToVariableMap(), coreBlock);
+      } else if (actor instanceof SrdagActor) {
+        generateFifoCall((SrdagActor) actor, coreBlock);
+      } else if (actor instanceof CommunicationActor) {
+        // TODO handle send/receive + enabler triggers
+        throw new PreesmRuntimeException("Unsupported actor [" + actor + "]");
+      } else {
+        throw new PreesmRuntimeException("Unsupported actor [" + actor + "]");
       }
     }
   }
 
+  private void generateFifoCall(SrdagActor actor, CoreBlock coreBlock) {
+    // Create the Fifo call and set basic property
+    final FifoCall fifoCall = CodegenModelUserFactory.eINSTANCE.createFifoCall();
+    fifoCall.setName(actor.getName());
+
+    final PortDirection dir;
+    final InitActor initActor;
+    if (actor instanceof InitActor) {
+      initActor = (InitActor) actor;
+      fifoCall.setOperation(FifoOperation.POP);
+      dir = PortDirection.OUTPUT;
+
+    } else if (actor instanceof EndActor) {
+      final EndActor endActor = (EndActor) actor;
+      initActor = (InitActor) endActor.getInitReference();
+      fifoCall.setOperation(FifoOperation.PUSH);
+      dir = PortDirection.INPUT;
+
+    } else {
+      throw new PreesmRuntimeException();
+    }
+
+    final Buffer buffer = (Buffer) memoryLinker.getPortToVariableMap().get(actor.getDataPort());
+    fifoCall.addParameter(buffer, dir);
+
+    final org.preesm.algorithm.memalloc.model.Buffer buffer2 = memAlloc.getDelayAllocations().get(initActor);
+    final Buffer delayBuffer = memoryLinker.getCodegenBuffer(buffer2);
+
+    fifoCall.setHeadBuffer(delayBuffer);
+    fifoCall.setBodyBuffer(null);
+
+    // Create the INIT call (only the first time the fifo is encountered)
+    // @farresti:
+    // the management of local/none and permanent delays is a bit dirty here.
+    // Ideally, this should be called only for permanent delay.
+    // local/none delays should only have a call to the init function, no need for pop/push.
+    // The case where the end vertex is alone should not be considered as it means that the tokens convoyed by the
+    // delay are discarded.
+    // Actually, only the permanent delays really need the pop/push mechanism.
+    if (fifoCall.getOperation().equals(FifoOperation.POP)) {
+      final FifoCall fifoInitCall = CodegenModelUserFactory.eINSTANCE.createFifoCall();
+      fifoInitCall.setOperation(FifoOperation.INIT);
+      fifoInitCall.setFifoHead(fifoCall);
+      fifoInitCall.setName(fifoCall.getName());
+      fifoInitCall.setHeadBuffer(fifoCall.getHeadBuffer());
+      fifoInitCall.setBodyBuffer(fifoCall.getBodyBuffer());
+      final PersistenceLevel level = initActor.getPersistenceLevel();
+      if (level == null || PersistenceLevel.PERMANENT.equals(level)) {
+        coreBlock.getInitBlock().getCodeElts().add(fifoInitCall);
+      } else {
+        coreBlock.getLoopBlock().getCodeElts().add(fifoInitCall);
+      }
+    }
+    // Add the Fifo call to the loop of its coreBlock
+    coreBlock.getLoopBlock().getCodeElts().add(fifoCall);
+  }
+
   private void generateSpecialActor(final SpecialActor actor, final Map<Port, Variable> portToVariable,
       final CoreBlock operatorBlock) {
-    if (actor instanceof InitActor || actor instanceof EndActor) {
-      // nothing
+    final SpecialCall specialCall = CodegenModelUserFactory.eINSTANCE.createSpecialCall();
+    specialCall.setName(actor.getName());
+
+    final Fifo uniqueFifo;
+    final Buffer lastBuffer;
+    if (actor instanceof JoinActor) {
+      specialCall.setType(SpecialType.JOIN);
+      uniqueFifo = actor.getDataOutputPorts().get(0).getFifo();
+      lastBuffer = this.memoryLinker.getCodegenBuffer(memAlloc.getFifoAllocations().get(uniqueFifo).getSourceBuffer());
+    } else if (actor instanceof ForkActor) {
+      specialCall.setType(SpecialType.FORK);
+      uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
+      lastBuffer = this.memoryLinker.getCodegenBuffer(memAlloc.getFifoAllocations().get(uniqueFifo).getTargetBuffer());
+    } else if (actor instanceof BroadcastActor) {
+      specialCall.setType(SpecialType.BROADCAST);
+      uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
+      lastBuffer = this.memoryLinker.getCodegenBuffer(memAlloc.getFifoAllocations().get(uniqueFifo).getTargetBuffer());
+    } else if (actor instanceof RoundBufferActor) {
+      specialCall.setType(SpecialType.ROUND_BUFFER);
+      uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
+      lastBuffer = this.memoryLinker.getCodegenBuffer(memAlloc.getFifoAllocations().get(uniqueFifo).getTargetBuffer());
     } else {
-      final SpecialCall f = CodegenModelUserFactory.eINSTANCE.createSpecialCall();
-      f.setName(actor.getName());
-
-      final Fifo uniqueFifo;
-      if (actor instanceof JoinActor) {
-        f.setType(SpecialType.JOIN);
-        uniqueFifo = actor.getDataOutputPorts().get(0).getFifo();
-      } else if (actor instanceof ForkActor) {
-        f.setType(SpecialType.FORK);
-        uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
-      } else if (actor instanceof BroadcastActor) {
-        f.setType(SpecialType.BROADCAST);
-        uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
-      } else if (actor instanceof RoundBufferActor) {
-        f.setType(SpecialType.ROUND_BUFFER);
-        uniqueFifo = actor.getDataInputPorts().get(0).getFifo();
-      } else {
-        throw new PreesmRuntimeException("special actor " + actor + " has an unknown special type");
-      }
-
-      final org.preesm.algorithm.memalloc.model.Buffer buffer = memAlloc.getAllocations().get(uniqueFifo);
-      final Buffer lastBuffer = allocation.btb.get(buffer);
-
-      // Add it to the specialCall
-      if (actor instanceof JoinActor) {
-        f.addOutputBuffer(lastBuffer);
-      } else {
-        f.addInputBuffer(lastBuffer);
-      }
-
-      operatorBlock.getLoopBlock().getCodeElts().add(f);
-
-      registerCallVariableToCoreBlock(operatorBlock, f);
+      throw new PreesmRuntimeException("special actor " + actor + " has an unknown special type");
     }
+
+    // Add it to the specialCall
+    if (actor instanceof JoinActor) {
+      specialCall.addOutputBuffer(lastBuffer);
+      actor.getDataInputPorts().stream().map(port -> ((Buffer) portToVariable.get(port)))
+          .forEach(specialCall::addInputBuffer);
+    } else {
+      specialCall.addInputBuffer(lastBuffer);
+      actor.getDataOutputPorts().stream().map(port -> ((Buffer) portToVariable.get(port)))
+          .forEach(specialCall::addOutputBuffer);
+    }
+
+    operatorBlock.getLoopBlock().getCodeElts().add(specialCall);
+
+    registerCallVariableToCoreBlock(operatorBlock, specialCall);
   }
 
   protected void registerCallVariableToCoreBlock(final CoreBlock operatorBlock, final Call call) {
@@ -448,25 +388,4 @@ public class CodegenModelGenerator2 {
       coreBlock.getLoopBlock().getCodeElts().add(loop);
     }
   }
-
-  private Map<Port, Variable> allocate() {
-    this.allocation = new AllocationVisitor();
-    this.allocation.doSwitch(this.memAlloc);
-    final Map<Port, Variable> portToVariable = this.allocation.portToVariable;
-
-    final EList<AbstractActor> allActors = this.algo.getAllActors();
-    for (final AbstractActor actor : allActors) {
-      for (final ConfigInputPort cip : actor.getConfigInputPorts()) {
-        final ISetter setter = cip.getIncomingDependency().getSetter();
-        if (setter instanceof Parameter) {
-          final long evaluate = ((Parameter) setter).getValueExpression().evaluate();
-          portToVariable.put(cip, CodegenModelUserFactory.eINSTANCE.createConstant(cip.getName(), evaluate));
-        } else {
-          throw new PreesmRuntimeException();
-        }
-      }
-    }
-    return portToVariable;
-  }
-
 }